+++ serve.d.ts
@@ -5,9 +5,6 @@ export interface JetpackServeOptions extends Omit<ResolveConfigInput, 'command'>
   command?: 'dev' | 'build'
 }
 
-export function serve(options?: JetpackServeOptions | ResolvedJetpackConfig): JetpackMiddleware
-export namespace serve {
-  function resolve(options?: JetpackServeOptions | ResolvedJetpackConfig): Promise<JetpackMiddleware>
-}
+export function serve(options?: JetpackServeOptions): JetpackMiddleware
 export function serveResolved(config: ResolvedJetpackConfig): JetpackMiddleware
 export default serve